requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosqlite==0.19.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
dnspython==2.6.1
sqlalchemy==2.0.27
alembic==1.13.1
email-validator==2.1.0.post1
greenlet==3.0.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
pyspf==2.0.14 